+++ CHANGES.md
@@ -1,3 +1,9 @@
+2018-08-09
+----------
+* Add `QmixPump.fill()` and `empty()` methods
+* `QmixPump` methods now raise on invalid volume and flow rate
+  parameters
+
 2018-08-07
 ----------
 * Fix reathedocs build